refactor: extract entity content loader and persistence from entity-store
- Extracted EntityContentLoader to handle content state and loading.
- Extracted EntityPersistenceService to handle debounced save coordination.
- Reduced entity-store.svelte.ts from 920 to 572 lines.
- Updated vite.config.ts to ensure new files run in node environment.
- Updated GOD_FILES_ANALYSIS.md report.
