+++ docs/GOD_FILES_ANALYSIS.md
@@ -4,18 +4,18 @@ This report identifies the top 10 potential "God Files" (files with excessive re
 
 ## Top 10 Largest Files (Excluding Tests & Generated Code)
 
-| Rank | File Path                                                   | Line Count      | Type                | Status   |
-| :--- | :---------------------------------------------------------- | :-------------- | :------------------ | :------- |
-| 1    | `apps/web/src/lib/stores/oracle.svelte.ts`                  | ~~1,484~~ 304   | Store (State/Logic) | ✅ FIXED |
-| 2    | `apps/web/src/lib/stores/vault.svelte.ts`                   | ~~1,381~~ 1,261 | Store (State/Logic) | ✅ FIXED |
-| 3    | `apps/web/src/lib/components/GraphView.svelte`              | ~~1,371~~ 561   | UI Component        | ✅ FIXED |
-| 4    | `apps/web/src/lib/components/modals/ZenModeModal.svelte`    | ~~1,058~~ 372   | UI Component        | ✅ FIXED |
-| 5    | `apps/web/src/lib/services/ai.ts`                           | ~~819~~ 1       | Service (API/Logic) | ✅ FIXED |
-| 6    | `apps/web/src/routes/(app)/+layout.svelte`                  | ~~795~~ 222     | UI Layout           | ✅ FIXED |
-| 7    | `apps/web/src/lib/components/map/MapView.svelte`            | ~~681~~ 448     | UI Component        | ✅ FIXED |
-| 8    | `packages/sync-engine/src/SyncService.ts`                   | 711             | Engine Core         | 🟡 SOON  |
-| 9    | `apps/web/src/lib/components/canvas/CanvasWorkspace.svelte` | 835             | UI Component        |          |
-| 10   | `apps/web/src/lib/components/oracle/ChatMessage.svelte`     | 629             | UI Component        |          |
+| Rank | File Path                                                   | Line Count    | Type                | Status   |
+| :--- | :---------------------------------------------------------- | :------------ | :------------------ | :------- |
+| 1    | `apps/web/src/lib/components/canvas/CanvasWorkspace.svelte` | 780           | UI Component        |          |
+| 2    | `apps/web/src/lib/stores/vault/entity-store.svelte.ts`      | ~~920~~ 572   | Store (State/Logic) | ✅ FIXED |
+| 3    | `apps/web/src/lib/components/GraphView.svelte`              | ~~1,371~~ 561 | UI Component        | ✅ FIXED |
+| 4    | `apps/web/src/lib/stores/vault.svelte.ts`                   | ~~1,381~~ 478 | Store (State/Logic) | ✅ FIXED |
+| 5    | `apps/web/src/lib/components/oracle/ChatMessage.svelte`     | 468           | UI Component        |          |
+| 6    | `apps/web/src/lib/components/map/MapView.svelte`            | ~~681~~ 448   | UI Component        | ✅ FIXED |
+| 7    | `apps/web/src/lib/components/modals/ZenModeModal.svelte`    | ~~1,058~~ 372 | UI Component        | ✅ FIXED |
+| 8    | `apps/web/src/lib/stores/oracle.svelte.ts`                  | ~~1,484~~ 304 | Store (State/Logic) | ✅ FIXED |
+| 9    | `packages/sync-engine/src/SyncService.ts`                   | ~~711~~ 242   | Engine Core         | ✅ FIXED |
+| 10   | `apps/web/src/routes/(app)/+layout.svelte`                  | ~~795~~ 222   | UI Layout           | ✅ FIXED |
 
 ---
 
@@ -31,7 +31,7 @@ This report identifies the top 10 potential "God Files" (files with excessive re
 
 **Status:** ✅ **COMPLETED (2026-03-11)**
 **Summary:** Refactored into `@codex/vault-engine` and specialized stores. Logic decoupled into `VaultRegistryStore`, `MapRegistryStore`, `CanvasRegistryStore`, and `VaultRepository`.
-**Outcome:** Reduced from ~1,400 lines to 1,261 lines. Improved data separation and persistence logic.
+**Outcome:** Reduced from ~1,400 lines to 478 lines. Improved data separation and persistence logic.
 
 ### 3. `GraphView.svelte` (Refactored)
 
@@ -55,25 +55,19 @@ This report identifies the top 10 potential "God Files" (files with excessive re
 
 **Status:** ✅ **COMPLETED (2026-03-22)**
 **Summary:** Restructured layout architecture using SvelteKit route groups. Created `(app)` route group for workspace routes, moved workspace routes into it, and stripped root layout to minimal shell.
-**Outcome:**
-
-- Root `+layout.svelte` reduced from 261 lines to 222 lines
-- Marketing pages no longer load workspace JS (stores, bootSystem, modals)
-- Clean separation between marketing routes (`(marketing)`) and workspace routes (`(app)`)
-- Build passes, all 152 tests pass
-
-**Changes:**
-
-- Created `routes/(app)/` route group with full workspace layout
-- Moved `+page.svelte`, `map/`, `canvas/`, `oracle/`, `help/`, `timeline/`, `import/` into `(app)`
-- Stripped root `+layout.svelte` to minimal HTML shell with SEO meta tags
-- Fixed relative imports in moved pages
+**Outcome:** Reduced from 795 lines to 222 lines.
 
 ### 7. `MapView.svelte` (Refactored)
 
 **Status:** ✅ **COMPLETED (2026-04-01)**
-**Summary:** Split into a thin composition shell with extracted loader, fog painter, pin popover, and pure interaction helpers. The remaining component now focuses on orchestration and rendering glue rather than owning every map concern directly.
-**Outcome:** Reduced from ~~681~~ 448 lines. The map-specific logic is now more testable and less tightly coupled.
+**Summary:** Split into a thin composition shell with extracted loader, fog painter, pin popover, and pure interaction helpers.
+**Outcome:** Reduced from ~~681~~ 448 lines.
+
+### 8. `entity-store.svelte.ts` (Refactored)
+
+**Status:** ✅ **COMPLETED (2026-04-27)**
+**Summary:** Extracted content loading and persistence coordination into `EntityContentLoader` and `EntityPersistenceService`.
+**Outcome:** Reduced from 920 lines to 572 lines. Improved modularity and separation of concerns.
 
 ---
